Introduce operational exploit module to the repo.
Contains a complete and separate ecosystem that allows the CSV output of a Greenbone OpenVAS scan to be automatically broken down into CVEs, searched against Metasploit's internal database of exploits and payloads, then executed. Can be modified to other CSV formats by adjusting the row[' '] titles. Includes small amounts of introductory documentation, and the basics of a system to build reports on.