Introduce jira_utils.py
As promised in the tradeshow, this is a script that allows the automation of Jira ticket creations. Currently, this pulls only from the CSV generated from the OpenVAS results, and in the interests of security should also require an account specifically made for the purpose of a least-access required security policy, being that it is needed only to create issues.